MindSPONGE/requirements.txt
numpy >= 1.17.0
biopython >= 1.79
parmed >= 3.4.3
pyyaml >= 5.4.1
dataclasses >= 0.6
glob2 >= 0.6
h5py >= 3.6.0
